// src/app/shared/models/request/EmployeeRequest.ts
export interface EmployeeRequest {
  id?: number;
  agencyId: number;
  firstName: string;
  middleName?: string;
  fatherLastName: string;
  motherLastName: string;
  statusId: number;
  positionId: number;
  birthDate: string;
  email: string;
  postalAddress: string;
  cityId: number;
  regionId: number;
  areaCode: string;
  comments?: string;
  userId?: string;
  isActive: boolean;
}